requirements.txt
pytest==9.0.1
appium-python-client==2.32.2
selenium==4.25.0
